Renamed BorderBuilderClass (fixed typo)
